Calculator hotfix
Changed some things in the calculator, changed the datatype from long to double. Added a button for decimal point that so far does nothing

+++ src/sample/Model.java
@@ -1,7 +1,7 @@
 package sample;
 
 public class Model {
-    public float Calculate(long number1, long number2, String operator) {
+    public double Calculate(double number1, double number2, String operator) {
 
         switch (operator) {
